+++ tests/rvm/mod.rs
@@ -1,4 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 mod compiler;
+#[cfg(all(feature = "std", feature = "yaml"))]
 mod rego;